r cmd check

+++ R/use_sfmisc.R
@@ -1,25 +1,52 @@
-#' Title
+#' Use sfmisc functions in current package
 #'
-#' @param x
-#'
-#' @return
+#' @return `TRUE` on success
 #' @export
 #'
-#' @examples
-use_sfmisc <- function(x){
-  assert_namespace("rprojroot")
+use_sfmisc <- function(){
+  stopifnot(requireNamespace("rprojroot"))
 
   infile <- system.file("utils", "utils-sfmisc.R", package = "sfmisc")
   stopifnot(file.exists(infile))
   outfile  <- rprojroot::find_package_root_file("R", "utils-sfmisc.R")
+  current_version <- utils::packageVersion("sfmisc")
 
-  unlink(outfile)
+  if (file.exists(outfile)){
+    message(sprintf(
+      "Replacing existing utils-sfmisc (v%s) with new version (v%s) at %s",
+      get_sfmisc_version(outfile),
+      current_version,
+      outfile
+    ))
+    unlink(outfile)
+  } else {
+    message(sprintf(
+      "Saving utils-sfmisc (v%s) to '%s'",
+      current_version,
+      outfile
+    ))
+
+  }
 
   write(
-    sprintf("# sfmisc utils %s\n\n\n\n", packageVersion("sfmisc")),
+    sprintf("# sfmisc utils %s\n\n\n\n", current_version),
     outfile
   )
 
 
+
   file.append(outfile, infile)
+}
+
+
+
+get_sfmisc_version <- function(x){
+  txt <- readLines(x)
+
+  if (length(txt) > 0){
+    regmatches(txt[[1]], regexpr("(\\d\\.\\d\\.\\d$)|(\\d\\.\\d\\.\\d\\.\\d*$)", txt[[1]]))
+  } else {
+    "*.*.*"
+  }
+
 }
